Add code field to GnuHealthTestCritearea and let lab interface script use it
We add code field to GnuHealthTestCritearea and let
doc/samples/interfaces/lab/gnuhealth_lab_interface.py to use this
field.

The reason to do this change is that:

name field is not suitable for interface stript, for it will changed
when user use different languages.

sequence field is not suitable for interface script too, for it may be
changed by user for sort reason, when it changed, interface script can
not find error. for example: when a criterea sequence is changed from
1 to 2 for sort reason. if interface script do not update, it will run
no error and push wrong value.

So we add new field code to do this job