bug #64934: Add image_crop_to_ratio to core and use it.
This feature is used to fix: bug #64934: Need to optimize formatting in reports

>  Pictures in the reports (Identification cards, newborn..) should keep the original aspect ratio, so they don't look distorted.


The reason to add image_crop_to-ratio function is that:
I can not find a way to keep ratio in fodt template file at the moment.
maybe others can find a better way to solve this problem.