gnuhealth-control: get_lang use msgcat to merge po files.
After gnuhealth-4.4, we use po-export tool to update po and pot files
of gnuhealth, po files will sync to gnuhealth-hg-dev. which means that
po items used in before gnuhealth version but removed in
gnuhealth-hg-dev will be removed and not sync to weblate to translate.

To solve this problem, we use msgcat to merge weblate's po files and
installed po files, but removed po items can not lost, but they will
be not updated with weblate, user should maintain them by manual.